N CUBE/backend/pdp/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Web3 / Blockchain
web3==6.18.0

# Cryptography
cryptography==42.0.7

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Async HTTP
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
